pipeline input selector component
